chore: URL과 URLRequest Extension 파일에 관련 메소드 작성 (#118)
Endpoint에 있던 extension에 작성한 로직을 별도 파일로 분리해줬습니다
